Adds java tcf worker
The java worker listens to new work order events on a blockchain,
filters them to only keep those related to its local wallet, downloads
the appropriate docker container, executes it locally and returns the
result to a smart contract.

All TCF effort (establishing the secure channel, signing the result
etc.) must be done in the container.

Signed-off-by: Anthony Simonet-Boulogne <anthony@iex.ec>